requirements.txt
pytest == 8.1.1
selene == 2.0.0rc9
allure-pytest == 2.13.5
python-dotenv == 1.0.1
requests == 2.32.3